refactor: move utility functions to separate files
