+++ src/core/packages/chrome/navigation/src/constants.ts
@@ -1,0 +1,15 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+/**
+ * The larger height of the primary menu item when the label is 2 lines
+ */
+export const EXPANDED_MENU_ITEM_HEIGHT = 67;
+export const COLLAPSED_MENU_ITEM_HEIGHT = 32;
+export const MAX_MENU_ITEMS = 10;
